Try to deal with builds for a different architecture
We apparently can have multiple architectures mixed into the same
build: one for srpm, one for archful packages, and one for noarch packages.
And it turns out that quite a few srpms have different BR depending
on where they are build. E.g. librtas in only available on ppc64le.
Previously, this script would strip the architecture of BR and try to
install the native version. But obviously this cannot work if the
package is not available for the native architecture. So try to skip
such packages and hope for the best. This might actually work in simple
cases, but is not a general solution. If we try to do a large-scale
rebuild, I think we might have to skip those cases where srpms were
built on a different architecture, because we cannot guess what the
BR actually were on our architecture. Or maybe we should only skip the
case where the the buildroot lists any srpms that are not available
on the rebuild architecture. But this obviously means that cases like
librtas would cause a failure if we were rebuilding on an architeture
that needs the extra package.