koji_rebuild: use tasks list to figure out the buildroot
For archful builds with a noarch output, koji will pick one of the noarch rpms
and drop the others. When rebuilding, we'll use this rpm and then possibly fail
the build when the buildroot content is different between architectures.
Instead, look at the list of tasks, and pick the task for our architecture,
even if the products were ultimately discarded.