koji_rebuild: do not treat builds with no rpms as a fatal error
optee_os-4.4.0-1.fc42 is an archful build, but all architectures except for
aarch64 do not produce any outputs. On aarch64, a noarch package is produced.
We treated this is a program bug, but it's really not. Let's just do
the comparison as usual and report that the package list was mismatched.